Move `runtimeExtensionsInput` to `/common/`
